refactor: migrate BunnyNet and Laravel Forge to new Integration pattern
Refactors the two remaining legacy host providers (BunnyNet and Laravel
Forge) to match the Integration + Capability_Module architecture used by
all other host integrations since 2.5.0.

Changes:
- Add inc/integrations/providers/bunnynet/class-bunnynet-integration.php
- Add inc/integrations/providers/bunnynet/class-bunnynet-domain-mapping.php
- Add inc/integrations/providers/laravel-forge/class-laravel-forge-integration.php
- Add inc/integrations/providers/laravel-forge/class-laravel-forge-domain-mapping.php
- Register both providers in Integration_Registry (core integrations + capabilities)
- Remove legacy Laravel_Forge_Host_Provider::get_instance() from Domain_Manager
- Delete inc/integrations/host-providers/class-bunnynet-host-provider.php
- Delete inc/integrations/host-providers/class-laravel-forge-host-provider.php

The new classes use get_credential() for credential access (supports both
wp-config.php constants and encrypted network options), and delegate API
calls from the domain mapping capability to the integration class.

Closes #449
